Dump of old site
This project uses another project called qdig to power its portfolio
section. Another side is currently under development. This site really
is just for historical purposes, but I may just spend the time updating
it if there is little else to do. If you see updates in this repo, it is
because I am incredibly bored and have absolutely nothing better to do.
;-)